resqnet-call/requirements.txt
fastapi
uvicorn
twilio
openai-whisper
google-generativeai
httpx
python-multipart
python-dotenv
scikit-learn
